pyproject.toml
[tool.poetry]
name = "replicating-simple-blackbox-attack"
version = "0.1.7"
description = "Replicating the code and results of the paper \"Simple Black-box Adversarial Attacks\""
authors = ["Eric Kolibacz"]
license = "MIT license"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
pandas = "^1.5.3"
mat4py = "^0.5.0"
json3 = "^1.0"
numpy = "^1.24.2"
matplotlib = "^3.7.1"
torchvision = "^0.15.1"
torch = "^2.0.0"
pillow = "^9.5.0"
ipykernel = "^6.22.0"
scipy = "^1.10.1"


[tool.poetry.group.code_style.dependencies]
pylint = "^2.17.2"
pycodestyle = "^2.10.0"
isort = "^5.12.0"
black = "^23.3.0"
nbqa = "^1.7.0"


[tool.poetry.group.testing.dependencies]
pytest = "^7.2.2"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true

[tool.pylint."MESSAGES CONTROL"]
max-line-length = 120
generated-members="torch.*"
